+++ tests/unit/Observers/UrlRewriteObserverTest.php
@@ -21,6 +21,7 @@
 namespace TechDivision\Import\Product\UrlRewrite\Observers;
 
 use PHPUnit\Framework\TestCase;
+use ReflectionClass;
 use TechDivision\Import\Utils\EntityStatus;
 use TechDivision\Import\Utils\StoreViewCodes;
 use TechDivision\Import\Product\Utils\VisibilityKeys;
@@ -74,8 +75,71 @@ class UrlRewriteObserverTest extends TestCase
         $this->observer = new UrlRewriteObserver($this->mockProductUrlRewriteProcessor);
     }
 
+    protected function createAndInvokeObserver($generateCategoryProductRewrites = true, $productCategoryIds = [], $category = [])
+    {
+        $observer = $this->getMockBuilder('TechDivision\Import\Product\UrlRewrite\Observers\UrlRewriteObserver')
+            ->setConstructorArgs([$this->mockProductUrlRewriteProcessor])
+            ->setMethods(['getGenerateCategoryProductRewritesOptionValue', 'isRootCategory'])
+            ->getMock();
+
+        $observer->expects($this->any())
+            ->method('getGenerateCategoryProductRewritesOptionValue')
+            ->willReturn($generateCategoryProductRewrites);
+
+        $observer->expects($this->any())
+            ->method('isRootCategory')
+            ->willReturn(false);
+
+        // prepare protected properties of observer
+        $reflection = new ReflectionClass('TechDivision\Import\Product\UrlRewrite\Observers\UrlRewriteObserver');
+        $property = $reflection->getProperty('productCategoryIds');
+        $property->setAccessible(true);
+        $property->setValue($observer, $productCategoryIds);
+
+        // invoke the method to test
+        $method = $reflection->getMethod('createProductCategoryRelation');
+        $method->setAccessible(true);
+        $method->invoke($observer, $category, true);
+
+        return $property->getValue($observer);
+    }
+
+    public function testCreateProductCategoryRelationWithChildCategoryAndSettingEnabled()
+    {
+
+        // initialize method arguments
+        $generateCategoryProductRewrites = true;
+        $productCategoryIds = ['2'];
+        $category = [
+            'entity_id' => '10',
+            'is_anchor' => '0'
+        ];
+
+        // create and invoke the partially mocked observer
+        $productCategoryIds = $this->createAndInvokeObserver($generateCategoryProductRewrites, $productCategoryIds, $category);
+
+        $this->assertSame($productCategoryIds, ['2', '10']);
+    }
+
+    public function testCreateProductCategoryRelationWithChildCategoryAndSettingDisabled()
+    {
+
+        // initialize method arguments
+        $generateCategoryProductRewrites = false;
+        $productCategoryIds = ['2'];
+        $category = [
+            'entity_id' => '10',
+            'is_anchor' => '0'
+        ];
+
+        // create and invoke the partially mocked observer
+        $productCategoryIds = $this->createAndInvokeObserver($generateCategoryProductRewrites, $productCategoryIds, $category);
+
+        $this->assertSame($productCategoryIds, ['2']);
+    }
+
     /**
-     * Test's the handle() method with a successfull URL rewrite persist.
+     * Test's the handle() method with a successful URL rewrite persist.
      *
      * @return void
      */
@@ -210,7 +274,7 @@ class UrlRewriteObserverTest extends TestCase
     }
 
     /**
-     * Test's the handle() method with a successfull URL rewrite persist when using the same categories.
+     * Test's the handle() method with a successful URL rewrite persist when using the same categories.
      *
      * @return void
      */
@@ -392,14 +456,13 @@ class UrlRewriteObserverTest extends TestCase
         $mockSubject->expects($this->any())
                     ->method('getRowStoreId')
                     ->willReturn($storeId = 1);
-        $mockSubject->expects($this->exactly(3))
+        $mockSubject->expects($this->any())
                     ->method('getCoreConfigData')
-                    ->withConsecutive(
-                        array(CoreConfigDataKeys::CATALOG_SEO_PRODUCT_URL_SUFFIX, '.html'),
-                        array(CoreConfigDataKeys::CATALOG_SEO_PRODUCT_URL_SUFFIX, '.html'),
-                        array(CoreConfigDataKeys::CATALOG_SEO_PRODUCT_URL_SUFFIX, '.html')
-                    )
-                    ->willReturnOnConsecutiveCalls('.html', '.html', '.html');
+                    ->will(
+                        $this->returnCallback(function ($arg1, $arg2) {
+                            return $arg2;
+                        })
+                    );
        $mockSubject->expects(($this->exactly(4)))
                     ->method('getImportAdapter')
                     ->willReturn($mockImportAdapter);